Updated executable and installer to v1.5.1
